Go binding: fix bug with R/O transaction destroyed before futures, add GC references for db/tx
In a previous PR the R/O transaction was optimized and the finalizer was removed; however this prevents using
any future, since transaction is already destroyed by the time future is attempted evaluation.

Additionally, a new issue is uncovered and addressed: the Go Database, Transaction and Future objects are all
garbage-collected independently, which is incorrect because Database and Transaction must not be garbage-collected
if any future has not yet been garbage-collected.